Move class to own header
